+++ src/main/java/home/EncryptPassword.java
@@ -1,3 +1,4 @@
+package home;
 import java.security.MessageDigest;
 import java.security.NoSuchAlgorithmException;
