+++ src/main/java/home/RegisterForm.java
@@ -1,3 +1,4 @@
+package home;
 
 
 import java.awt.BorderLayout;
