+++ src/main/java/home/home.java
@@ -1,3 +1,4 @@
+package home;
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
 import java.awt.Image;
