+++ src/main/java/home/pages/member/AllFlightsMemb.java
@@ -1,3 +1,4 @@
+package home.pages.member;
 import java.awt.BorderLayout;
 import java.awt.EventQueue;
 import java.sql.Connection;
